% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{create_desolve_times}
\alias{create_desolve_times}
\title{Create a deSolve-compatible times sequence}
\usage{
create_desolve_times(model, duration, step, unit = "h")
}
\arguments{
\item{model}{An object of class \code{sbmlModel}.}

\item{duration}{A number describing the total duration of the simulation.}

\item{step}{A number describing the resolution in terms of step size.}

\item{unit}{An optional character string describing the time unit of the
specified duration and step length. Timings are aligned with the
time unit of the \code{model}. Current options are "h" for hours (default),
"d" for days, or "model" when assuming the specified duration and step length
are considered in line with the time unit of the \code{model}.}
}
\value{
A \code{vector} of simulation output times formatted for use as the
\code{times} argument in \code{\link[deSolve]{ode}}.
}
\description{
Creates a \code{deSolve} compliant \code{times} sequence that can be used as
the \code{times} argument in \code{\link[deSolve]{ode}}, specifying the times for
which simulation output is wanted.
}
\examples{
model_file <- system.file("extdata/", "simple_oral.sbml", package = "sbmlpbk")
model <- load_sbml(model_file)

times <- create_desolve_times(
  model,
  duration = 10, # 10 days
  step = 1/24,   # every hour
  unit = 'd'     # times in days
)

head(times)

}
\seealso{
\code{\link[deSolve]{ode}}
}
