% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{load_functions}
\alias{load_functions}
\title{Loads the functions from the SBML model}
\usage{
load_functions(model, envir = parent.frame())
}
\arguments{
\item{model}{SBML model S3 class.}

\item{envir}{Environment in which to load the functions. Default the
the environment from which the function is called.}
}
\value{
An object of class \code{sbmlModel}.
}
\description{
Loads the functions from the SBML model
}
