% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_params.R
\name{load_params}
\alias{load_params}
\title{Load PBK model parameter values from a CSV}
\usage{
load_params(model, filename, param_instance, ...)
}
\arguments{
\item{model}{An object of class \code{sbmlModel} for which the parametrisations should be loaded.}

\item{filename}{Path to the CSV file containing parameters.}

\item{param_instance}{The instance ID used to filter parameters. When not specified, it is assumed that the file contains only one parametrisation.}

\item{...}{Additional arguments passed to \code{read.csv}.}
}
\value{
A named numeric vector with parameter values.
}
\description{
This function reads a CSV file of parameter values and returns a named
numeric vector. Only rows matching the specified \code{idInstance} are used.
}
\details{
The input CSV file must contain the following columns:
\describe{
\item{\code{idModelInstance}}{A character string identifying the model instance. Used for filtering.}
\item{\code{Parameter}}{The name of the parameter to be loaded. These will become the names of the output vector.}
\item{\code{Value}}{The numeric value associated with each parameter.}
}
}
\examples{
model_file <- system.file("extdata/", "simple_oral.sbml", package = "sbmlpbk")
model <- load_sbml(model_file)
params_file <- system.file("extdata/", "simple_oral_params.csv", package = "sbmlpbk")
params <- load_params(model, params_file, 'simple_PARAM')
head(params)

# Example CSV structure:
# idModelInstance,Parameter,Value
# "model_1",a,0.5
# "model_2",a,1.0
# "model_2",b,2.5

}
