% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_sbml.R
\name{load_sbml}
\alias{load_sbml}
\title{Load SBML model}
\usage{
load_sbml(sbml_file)
}
\arguments{
\item{sbml_file}{File path of the SBML file.}
}
\value{
An object of class \code{sbmlModel}.
}
\description{
Load SBML model
}
\examples{
file <- system.file("extdata/", "simple_oral.sbml", package = "sbmlpbk")
model <- load_sbml(file)

}
