% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_model.R
\name{print.sbmlModel}
\alias{print.sbmlModel}
\title{Print an \code{sbmlModel} Object}
\usage{
\method{print}{sbmlModel}(x, ...)
}
\arguments{
\item{x}{An object of class \code{sbmlModel}.}

\item{...}{Additional arguments. The following argument can be passed via \code{...}:
\describe{
\item{\code{type}}{A character string, either \code{"summary"} (default) or \code{"equations"}.
Controls whether a high-level summary or the model's differential equations are printed.}
}}
}
\description{
This is a print method for objects of class \code{sbmlModel}. It supports multiple
output modes depending on the \code{type} argument.
}
\examples{
file <- system.file("extdata/", "simple_oral.sbml", package = "sbmlpbk")
model <- load_sbml(file)
print(model, type = "summary")
print(model, type = "equations")

}
