% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_model.R
\name{summary.sbmlModel}
\alias{summary.sbmlModel}
\title{Summary function for the class \code{sbmlModel}}
\usage{
\method{summary}{sbmlModel}(object, ...)
}
\arguments{
\item{object}{An object of class \code{sbmlModel}.}

\item{...}{Not used.}
}
\value{
A list with summary items.
}
\description{
Gives a summary for an object of S3 class \code{sbmlModel}.
}
